Support large number of calls

+++ src/alltoall/mpi_alltoall.c
@@ -489,7 +489,7 @@ static int insert_sendrecv_data(int *sbuf, int *rbuf, int size, int sendtype_siz
 			if (temp->count >= temp->max_calls)
 			{
 				temp->max_calls = temp->max_calls * 2;
-				temp->list_calls = (int *)realloc(temp->list_calls, temp->max_calls * sizeof(int));
+				temp->list_calls = (uint64_t *)realloc(temp->list_calls, temp->max_calls * sizeof(uint64_t));
 				assert(temp->list_calls);
 			}
 			temp->list_calls[temp->count] = avCalls; // Note: count starts at 1, not 0
@@ -510,7 +510,7 @@ static int insert_sendrecv_data(int *sbuf, int *rbuf, int size, int sendtype_siz
 
 	newNode->size = size;
 	newNode->count = 1;
-	newNode->list_calls = (int *)malloc(DEFAULT_TRACKED_CALLS * sizeof(int));
+	newNode->list_calls = (uint64_t *)malloc(DEFAULT_TRACKED_CALLS * sizeof(uint64_t));
 	assert(newNode->list_calls);
 	newNode->max_calls = DEFAULT_TRACKED_CALLS;
 	// We have at most <size> different counts (one per rank) and we just allocate pointers of pointers here, not much space used  //TODO adapt to counts for alltoall (cf alltoallv)
@@ -611,9 +611,9 @@ static void _save_patterns(FILE *fh, avPattern_t *p, char *ctx)
 	while (ptr != NULL)
 	{
 #if COMMSIZE_BASED_PATTERNS || TRACK_PATTERNS_ON_CALL_BASIS
-		fprintf(fh, "During %d alltoall calls, %d ranks %s %d other ranks; comm size: %d\n", ptr->n_calls, ptr->n_ranks, ctx, ptr->n_peers, ptr->comm_size);
+		fprintf(fh, "During %"PRIu64" alltoall calls, %d ranks %s %d other ranks; comm size: %d\n", ptr->n_calls, ptr->n_ranks, ctx, ptr->n_peers, ptr->comm_size);
 #else
-		fprintf(fh, "During %d alltoall calls, %d ranks %s %d other ranks\n", ptr->n_calls, ptr->n_ranks, ctx, ptr->n_peers);
+		fprintf(fh, "During %"PRIu64" alltoall calls, %d ranks %s %d other ranks\n", ptr->n_calls, ptr->n_ranks, ctx, ptr->n_peers);
 #endif // COMMSIZE_BASED_PATTERNS
 		ptr = ptr->next;
 	}
@@ -642,7 +642,7 @@ static void save_call_patterns(int uniqueID)
 	avCallPattern_t *ptr = call_patterns;
 	while (ptr != NULL)
 	{
-		fprintf(fh, "For %d call(s):\n", ptr->n_calls);
+		fprintf(fh, "For %"PRIu64" call(s):\n", ptr->n_calls);
 		_save_patterns(fh, ptr->spatterns, "sent to");
 		_save_patterns(fh, ptr->rpatterns, "recv'd from");
 		ptr = ptr->next;
@@ -983,7 +983,7 @@ static int _commit_data()
 	return 0;
 }
 
-static caller_info_t *create_new_caller_info(char *caller, int n_call)
+static caller_info_t *create_new_caller_info(char *caller, uint64_t n_call)
 {
 	caller_info_t *new_info = malloc(sizeof(caller_info_t));
 	assert(new_info);
@@ -996,7 +996,7 @@ static caller_info_t *create_new_caller_info(char *caller, int n_call)
 	return new_info;
 }
 
-static int insert_caller_data(char **trace, size_t size, int n_call, int world_rank)
+static int insert_caller_data(char **trace, size_t size, uint64_t n_call, int world_rank)
 {
 	char *filename = NULL;
 	char *target_dir = NULL;
@@ -1010,7 +1010,7 @@ static int insert_caller_data(char **trace, size_t size, int n_call, int world_r
 	{
 		target_dir = strdup("backtraces");
 	}
-	_asprintf(filename, rc, "%s/backtrace_rank%d_call%d.md", target_dir, world_rank, n_call);
+	_asprintf(filename, rc, "%s/backtrace_rank%d_call%"PRIu64".md", target_dir, world_rank, n_call);
 	assert(rc > 0);
 
 	// Make sure the target directory exists
@@ -1037,7 +1037,7 @@ static int insert_caller_data(char **trace, size_t size, int n_call, int world_r
 	free(filename);
 }
 
-static void save_times(double *times, int comm_size, int n_call)
+static void save_times(double *times, int comm_size, uint64_t n_call)
 {
 	char *filename = NULL;
 	int i;
@@ -1046,22 +1046,22 @@ static void save_times(double *times, int comm_size, int n_call)
 #ifdef ENABLE_A2A_TIMINGS
 	if (getenv(OUTPUT_DIR_ENVVAR))
 	{
-		_asprintf(filename, rc, "%s/a2a_execution_times.rank%d_call%d.md", getenv(OUTPUT_DIR_ENVVAR), world_rank, n_call);
+		_asprintf(filename, rc, "%s/a2a_execution_times.rank%d_call%"PRIu64".md", getenv(OUTPUT_DIR_ENVVAR), world_rank, n_call);
 	}
 	else
 	{
-		_asprintf(filename, rc, "a2a_execution_times.rank%d_call%d.md", world_rank, n_call);
+		_asprintf(filename, rc, "a2a_execution_times.rank%d_call%"PRIu64".md", world_rank, n_call);
 	}
 #endif // ENABLE_A2A_TIMINGS
 
 #ifdef ENABLE_LATE_ARRIVAL_TIMING
 	if (getenv(OUTPUT_DIR_ENVVAR))
 	{
-		_asprintf(filename, rc, "%s/late_arrival_times.rank%d_call%d.md", getenv(OUTPUT_DIR_ENVVAR), world_rank, n_call);
+		_asprintf(filename, rc, "%s/late_arrival_times.rank%d_call%"PRIu64".md", getenv(OUTPUT_DIR_ENVVAR), world_rank, n_call);
 	}
 	else
 	{
-		_asprintf(filename, rc, "late_arrival_times.rank%d_call%d.md", world_rank, n_call);
+		_asprintf(filename, rc, "late_arrival_times.rank%d_call%"PRIu64".md", world_rank, n_call);
 	}
 #endif // ENABLE_LATE_ARRIVAL_TIMING
 	assert(rc > 0);
